+++ envdiff/__init__.py
@@ -1,4 +1,4 @@
 __project__ = 'env-diff'
-__version__ = '0.1'
+__version__ = '0.2'
 
 VERSION = "{0} v{1}".format(__project__, __version__)