+++ CHANGELOG.md
@@ -3,3 +3,8 @@
 ## 0.1 (2017-03-30)
 
 - Initial alpha release.
+
+## 0.2 (2017-03-30)
+
+- Added detection of source variables matching `export FOO=bar`.
+- Added Markdown report generation.